pathway_pipeline/requirements.txt
pathway
pandas
numpy
geopy
psycopg2-binary
python-dotenv
